apps/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "python-dotenv",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "fastapi",
    "uvicorn",
    "redis",
    "rq",
    "langchain-ollama>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-tavily>=0.2.13",
    "orjson>=3.11.4",
    "pika>=1.3.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
